@@ -2,8 +2,8 @@ package com.mylaputa.beleco;
 
 import android.app.WallpaperManager;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
 
 /**
  * Created by dklap on 1/22/2017.
